+++ firmware/controllers/math/firing_order.cpp
@@ -12,6 +12,7 @@ static const uint8_t order_1_THEN_3_THEN_4_THEN2[] = { 1, 3, 4, 2 };
 static const uint8_t order_1_THEN_2_THEN_4_THEN3[] = { 1, 2, 4, 3 };
 static const uint8_t order_1_THEN_3_THEN_2_THEN4[] = { 1, 3, 2, 4 };
 static const uint8_t order_1_THEN_4_THEN_3_THEN2[] = { 1, 4, 3, 2 };
+static const uint8_t order_1_THEN_2_THEN_3_THEN4[] = { 1, 2, 3, 4 };
 
 // 5 cylinder
 static const uint8_t order_1_2_4_5_3[] = {1, 2, 4, 5, 3};
@@ -71,6 +72,7 @@ static size_t getFiringOrderLength() {
 	case FO_1_2_4_3:
 	case FO_1_3_2_4:
 	case FO_1_4_3_2:
+	case FO_1_2_3_4:
 		return 4;
 // 5 cylinder
 	case FO_1_2_4_5_3:
@@ -146,6 +148,8 @@ static const uint8_t* getFiringOrderTable() {
 		return order_1_THEN_3_THEN_2_THEN4;
 	case FO_1_4_3_2:
 		return order_1_THEN_4_THEN_3_THEN2;
+	case FO_1_2_3_4:
+		return order_1_THEN_2_THEN_3_THEN4;
 // 5 cylinder
 	case FO_1_2_4_5_3:
 		return order_1_2_4_5_3;