+++ firmware/CHANGELOG.md
@@ -15,6 +15,7 @@ Release template (copy/paste this for new release):
 
 ### Added
  - MAP Prediction AE Improvements #9482
+ - 1-2-3-4 firing order for 982 honda sabre 750 reverse rotation V4
 
 ### Fixed
  -